Move ConfirmDeletionModal to src/components/elements/samples/measurementsTab/
