+++ app/packs/src/components/elements/samples/measurementsTab/ConfirmDeletionModal.js
